+++ internal/controllers/gateway/route_predicates.go
@@ -67,9 +67,8 @@ func IsRouteAttachedToReconciledGateway[routeT gatewayapi.RouteT](
 		if parentRef.Group != nil {
 			group = string(*parentRef.Group)
 		}
-
-		switch {
-		case kind == "Gateway" && group == gatewayapi.GroupVersion.Group:
+		// Check the parent gateway if the parentRef points to a gateway that is possible to be controlled by KIC.
+		if kind == "Gateway" && group == gatewayapi.GroupVersion.Group {
 			var gateway gatewayapi.Gateway
 			err := cl.Get(context.Background(), k8stypes.NamespacedName{Namespace: namespace, Name: string(parentRef.Name)}, &gateway)
 			if err != nil {
@@ -87,12 +86,6 @@ func IsRouteAttachedToReconciledGateway[routeT gatewayapi.RouteT](
 			if isGatewayClassControlled(&gatewayClass) {
 				return true
 			}
-		default:
-			log.Error(
-				fmt.Errorf("unsupported parentRef kind %s and group %s", kind, group),
-				"Got an unexpected kind and group when checking route's parentRefs",
-			)
-			return false
 		}
 	}